2539: move secrets overlay into secrets folder. Only pass inject prop to SecretsList.
