+++ src/overlays/components/OverlayController.tsx
@@ -35,7 +35,7 @@ import AutoRefreshOverlay from 'src/dashboards/components/AutoRefreshOverlay'
 import CellCloneOverlay from 'src/shared/components/cells/CellCloneOverlay'
 import CustomApiTokenOverlay from 'src/authorizations/components/CustomApiTokenOverlay'
 import DisplayTokenOverlay from 'src/authorizations/components/DisplayTokenOverlay'
-import CreateSecretOverlay from 'src/shared/components/secrets/CreateSecretOverlay'
+import CreateSecretOverlay from 'src/secrets/components/CreateSecret/CreateSecretOverlay'
 
 // Actions
 import {dismissOverlay} from 'src/overlays/actions/overlays'
